Find a maximum likelihood fit of the POUMM model

Find a maximum likelihood fit of the POUMM model

maxLikPOUMMGivenTreeVTips(loglik, pruneInfo, parLower, parUpper,
  parInitML = NULL, control = list(factr = 1e+08, fnscale = -1),
  verbose = FALSE, debug = FALSE, ...)

Arguments

loglik
function(par, memo, parFixedNoAlpha)

pruneInfo
a list-object returned by the pruneTree(tree, z) function.

parLower, parUpper
Two named numeric vectors indicating the boundaries of the search region. Default values are parLower = c(alpha = 0, theta = 0, sigma = 0, sigmae = 0) and parUpper = c(alpha = 100, theta = 10, sigma = 20, sigmae = 10).

parInitML
A named vector (like parLower and parUpper) or a list of such vectors - starting points for optim.

control
List of parameters passed on to optim, default list(factr = 1e8, fnscale = -1), see ?optim.

verbose
A logical indicating whether to print informative messages on the standard output.

debug
A logical indicating whether to print debug messages (currently not implemented).

...
currently not used.

Value

a list containing an element par and an element value as well as the parameters passed