Summarize the results of a POUMM-fit

Summarize the results of a POUMM-fit

# S3 method for POUMM
summary(object, ..., startMCMC = NA, endMCMC = NA,
  thinMCMC = 1000, stats = statistics(object), mode = c("short",
  "long", "expert"))

Arguments

object
a POUMM object returned by POUMM-function (see ?POUMM).

...
Not used, but declared for consistency with the generic method summary.

startMCMC, endMCMC
integers indicating the range of the MCMC chains to be used for the analysis (excluding the initial warm-up phase)

thinMCMC
thinning interval of the MCMC chain to avoid strong autocorrelation between sampled elements;

stats
a named list of functions of the form function(par) number , which are called for each sample of each mcmc chain in object. Defaults to a call of statistics(object) returning a list of statistics functions relevant for the object. See also statistics.

mode
a character indicating the desired format of the returned summary as follows: 'short' - a data.table with the ML and MCMC estimates of heritability, model parameters, root-value and other statistics. 'long' - same information as in 'short' but including also the samples, which can be convenient for